refactor(spec): split temporal-guides into interval and ticks
Separate portable interval grammar/label contracts from calendar tick
generation, keep temporal-guides.ts as a stable facade, and split the
interval test suite into grammar vs stepping files.
