+++ .changeset/spec-temporal-guides-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Split temporal guides into interval grammar/labels and tick generation behind a stable temporal-guides facade. Public package exports and interval tick behavior are unchanged.
